Refactor quiz logic into separate component
